requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bcrypt==3.2.2
Django==4.0.5
django-braces>=1.15.0
django-crispy-forms>=1.14.0
django-ckeditor>=6.5.1
djangorestframework>=3.13.1
requests>=2.28.1
python-dotenv==0.21.0